+++ system/dbt/models/marts/dim_demo_customers.sql
@@ -1,0 +1,7 @@
+-- Marts model: customer dimension.
+
+select
+    customer_id,
+    customer_name,
+    signup_at
+from {{ ref('stg_demo_customers') }}